+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Adding the `setReactions` method to control the selector frame when the reactions change
 - Fixing reaction label text opacity
 - Fixing button selector when there is only 1 reaction
+- Fixing selector display alignment when the button is pushed
 
 ## [Version 1.0.2](https://github.com/yannickl/Reactions/releases/tag/1.0.2)
 *Released on 2015-10-11.*